Solaris warnings fixes.
git-svn-id: file:///home/yoann/SVN/libprelude/trunk@10082 09c5ec92-17d4-0310-903a-819935f44dba

+++ prelude-admin/prelude-admin.c
@@ -1973,13 +1973,13 @@ static int list_cmd(int argc, char **argv)
 
                 pw = getpwuid(st.st_uid);
                 if ( ! pw )
-                        snprintf(uidbuf, sizeof(uidbuf), "%d", st.st_uid);
+                        snprintf(uidbuf, sizeof(uidbuf), "%d", (int) st.st_uid);
                 else
                         snprintf(uidbuf, sizeof(uidbuf), "%s", pw->pw_name);
 
                 gr = getgrgid(st.st_gid);
                 if ( ! gr )
-                        snprintf(gidbuf, sizeof(gidbuf), "%d", st.st_gid);
+                        snprintf(gidbuf, sizeof(gidbuf), "%d", (int) st.st_gid);
                 else
                         snprintf(gidbuf, sizeof(gidbuf), "%s", gr->gr_name);
